refactor(oh7): Lot 2b — move init_db/path-resolution to bin db module, storage now core-free (#252)
